requirements.txt
# Core Django + DRF
Django>=4.2,<5.0
djangorestframework>=3.14.0

# JWT Authentication
djangorestframework-simplejwt>=5.2.2

# OTP Support
django-otp>=1.2.0

# Role-Based Permissions (custom logic – no third-party needed, but adding this for completeness)
django-guardian>=2.4.0  # Optional if you want object-level permissions

# CORS for API access from frontend
django-cors-headers>=4.0.0

# Environment variable management
python-decouple>=3.8

# For better password hashing
bcrypt>=4.0.1

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
coverage>=7.3.0

# Optional: If using PostgreSQL
psycopg2-binary>=2.9.9



python-decouple

